Learning-gen-ai-course/pyproject.toml
[project]
name = "learning-gen-ai-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.8",
    "duckduckgo-search>=8.1.1",
    "faiss-cpu>=1.13.2",
    "google-generativeai>=0.8.6",
    "groq>=0.37.1",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-google-genai>=4.2.2",
    "langchain-groq>=1.1.2",
    "langgraph>=1.1.8",
    "pydantic>=2.13.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=15.0.0",
    "tenacity>=9.1.4",
    "tiktoken>=0.12.0",
    "typer>=0.24.1",
]
